Refactor API, needs unit tests #81 #82
