+++ classes/testing/api/rest/route.php
@@ -14,6 +14,7 @@ namespace ingot\testing\api\rest;
 
 use ingot\permissions;
 use ingot\testing\ingot;
+use ingot\testing\types;
 
 abstract class route  {
 
@@ -25,7 +26,6 @@ abstract class route  {
 	 * @var string
 	 */
 	protected $what;
-
 
 
 	/**
@@ -410,4 +410,37 @@ abstract class route  {
 
 		return $base;
 	}
+
+	/**
+	 * Validate test type
+	 *
+	 * @since 0.2.0
+	 *
+	 * @param string $value
+	 *
+	 * @return bool
+	 */
+	public function allowed_type( $value ) {
+		return in_array( $value, types::allowed_types() );
+	}
+
+	/**
+	 * Validate sub_type click type
+	 *
+	 * @since 0.4.0
+	 *
+	 * @param $value
+	 *
+	 * @return bool
+	 */
+	public function allowed_sub_tupe( $value, $request ) {
+		$type = $request->get_param( 'type' );
+		if( 'click' == $type ) {
+			return in_array( $value, types::allowed_click_types() );
+		}else{
+			return in_array( $value, types::allowed_price_types() );
+		}
+
+	}
+
 }
